+++ tests/architecture/NativeUniversalBalance/functions/DepositNative.t.sol
@@ -18,7 +18,7 @@ contract DepositNativeTest is TestBaseNativeUniversalBalance {
     function test_depositNative_fail_whenHasNoEnoughETH_fuzzed(
         uint256 amount
     ) public {
-        vm.assume(amount < type(uint216).max);
+        vm.assume(amount < type(uint256).max);
 
         deal(user1, amount);
 
@@ -54,7 +54,7 @@ contract DepositNativeTest is TestBaseNativeUniversalBalance {
     function test_depositNative_success_withLend_fuzzed(
         uint256 amount
     ) public {
-        vm.assume(0 < amount && amount < type(uint216).max / _ONE);
+        vm.assume(0 < amount && amount < type(uint256).max / _ONE);
 
         deal(user1, amount);
 
@@ -89,7 +89,7 @@ contract DepositNativeTest is TestBaseNativeUniversalBalance {
     function test_depositNative_success_withoutLend_fuzzed(
         uint256 amount
     ) public {
-        vm.assume(0 < amount && amount < type(uint216).max / _ONE);
+        vm.assume(0 < amount && amount < type(uint256).max / _ONE);
 
         deal(user1, amount);
